feat: move collection types to own files
