+++ mapproxy/request/base.py
@@ -19,149 +19,8 @@ Service requests (parsing, handling, etc).
 from urllib.parse import parse_qsl, quote
 
 from mapproxy.util.py import cached_property
-
-
-class NoCaseMultiDict(dict):
-    """
-    This is a dictionary that allows case insensitive access to values.
-
-    >>> d = NoCaseMultiDict([('A', 'b'), ('a', 'c'), ('B', 'f'), ('c', 'x'), ('c', 'y'), ('c', 'z')])
-    >>> d['a']
-    'b'
-    >>> d.get_all('a')
-    ['b', 'c']
-    >>> 'a' in d and 'b' in d
-    True
-    """
-
-    def _gen_dict(self, mapping=()):
-        """A `NoCaseMultiDict` can be constructed from an iterable of
-        ``(key, value)`` tuples or a dict.
-        """
-        tmp = {}
-        if isinstance(mapping, NoCaseMultiDict):
-            for key, value in mapping.iteritems():
-                tmp.setdefault(key.lower(), (key, []))[1].extend(value)
-        else:
-            if isinstance(mapping, dict):
-                itr = mapping.items()
-            else:
-                itr = iter(mapping)
-            for key, value in itr:
-                tmp.setdefault(key.lower(), (key, []))[1].append(value)
-        return tmp
-
-    def __init__(self, mapping=()):
-        """A `NoCaseMultiDict` can be constructed from an iterable of
-        ``(key, value)`` tuples or a dict.
-        """
-        super().__init__(self._gen_dict(mapping))
-
-    def update(self, mapping=(), append=False):
-        """A `NoCaseMultiDict` can be updated from an iterable of
-        ``(key, value)`` tuples or a dict.
-        """
-        for _, (key, values) in self._gen_dict(mapping).items():
-            self.set(key, values, append=append, unpack=True)
-
-    def __getitem__(self, key):
-        """
-        Return the first data value for this key.
-
-        :raise KeyError: if the key does not exist
-        """
-        if key in self:
-            return dict.__getitem__(self, key.lower())[1][0]
-        raise KeyError(key)
-
-    def __setitem__(self, key, value):
-        dict.setdefault(self, key.lower(), (key, []))[1][:] = [value]
-
-    def __delitem__(self, key):
-        dict.__delitem__(self, key.lower())
-
-    def __contains__(self, key):
-        return dict.__contains__(self, key.lower())
-
-    def __getstate__(self):
-        data = []
-        for key, values in self.iteritems():
-            for v in values:
-                data.append((key, v))
-        return data
-
-    def __setstate__(self, data):
-        self.__init__(data)
-
-    def get(self, key, default=None, type_func=None):
-        """Return the default value if the requested data doesn't exist.
-        If `type_func` is provided and is a callable it should convert the value,
-        return it or raise a `ValueError` if that is not possible.  In this
-        case the function will return the default as if the value was not
-        found.
-
-        Example:
-
-        >>> d = NoCaseMultiDict(dict(foo='42', bar='blub'))
-        >>> d.get('foo', type_func=int)
-        42
-        >>> d.get('bar', -1, type_func=int)
-        -1
-        """
-        try:
-            rv = self[key]
-            if type_func is not None:
-                rv = type_func(rv)
-        except (KeyError, ValueError):
-            rv = default
-        return rv
-
-    def get_all(self, key):
-        """
-        Return all values for the key as a list. Returns an empty list, if
-        the key doesn't exist.
-        """
-        if key in self:
-            return dict.__getitem__(self, key.lower())[1]
-        else:
-            return []
-
-    def set(self, key, value, append=False, unpack=False):
-        """
-        Set a `value` for the `key`. If `append` is ``True`` the value will be added
-        to other values for this `key`.
-
-        If `unpack` is True, `value` will be unpacked and each item will be added.
-        """
-        if key in self:
-            if not append:
-                dict.__getitem__(self, key.lower())[1][:] = []
-        else:
-            dict.__setitem__(self, key.lower(), (key, []))
-        if unpack:
-            for v in value:
-                dict.__getitem__(self, key.lower())[1].append(v)
-        else:
-            dict.__getitem__(self, key.lower())[1].append(value)
-
-    def iteritems(self):
-        """
-        Iterates over all keys and values.
-        """
-        for _, (key, values) in dict.items(self):
-            yield key, values
-
-    def copy(self):
-        """
-        Returns a copy of this object.
-        """
-        return self.__class__(self)
-
-    def __repr__(self):
-        tmp = []
-        for key, values in self.iteritems():
-            tmp.append((key, values))
-        return '%s(%r)' % (self.__class__.__name__, tmp)
+from mapproxy.request.no_case_multi_dict import NoCaseMultiDict
+from mapproxy.request.request_params import RequestParams
 
 
 def url_decode(qs, charset='utf-8', decode_keys=False, include_empty=True,
@@ -280,111 +139,6 @@ class Request(object):
                 + quote(self.environ.get('SCRIPT_NAME', '').rstrip('/'))
                 + quote(self.environ.get('PATH_INFO', ''))
                 )
-
-
-class RequestParams(object):
-    """
-    This class represents key-value request parameters. It allows case-insensitive
-    access to all keys. Multiple values for a single key will be concatenated
-    (eg. to ``layers=foo&layers=bar`` becomes ``layers: foo,bar``).
-
-    All values can be accessed as a property.
-
-    :param param: A dict or ``NoCaseMultiDict``.
-    """
-    params = None
-
-    def __init__(self, param=None):
-        self.delimiter = ','
-
-        if param is None:
-            self.params = NoCaseMultiDict()
-        else:
-            self.params = NoCaseMultiDict(param)
-
-    def __str__(self):
-        return self.query_string
-
-    def get(self, key, default=None, type_func=None):
-        """
-        Returns the value for `key` or the `default`. `type_func` is called on the
-        value to alter the value (e.g. use ``type_func=int`` to get ints).
-        """
-        return self.params.get(key, default, type_func)
-
-    def set(self, key, value, append=False, unpack=False):
-        """
-        Set a `value` for the `key`. If `append` is ``True`` the value will be added
-        to other values for this `key`.
-
-        If `unpack` is True, `value` will be unpacked and each item will be added.
-        """
-        self.params.set(key, value, append=append, unpack=unpack)
-
-    def update(self, mapping=(), append=False):
-        """
-        Update internal request parameters from an iterable of ``(key, value)``
-        tuples or a dict.
-
-        If `append` is ``True`` the value will be added to other values for
-        this `key`.
-        """
-        self.params.update(mapping, append=append)
-
-    def __getattr__(self, name):
-        if name in self:
-            return self[name]
-        else:
-            raise AttributeError("'%s' object has no attribute '%s" %
-                                 (self.__class__.__name__, name))
-
-    def __getitem__(self, key):
-        return self.delimiter.join(map(str, self.params.get_all(key)))
-
-    def __setitem__(self, key, value):
-        """
-        Set `value` for the `key`. Does not append values (see ``MapRequest.set``).
-        """
-        self.set(key, value)
-
-    def __delitem__(self, key):
-        if key in self:
-            del self.params[key]
-
-    def iteritems(self):
-        for key, values in self.params.iteritems():
-            yield key, self.delimiter.join((str(x) for x in values))
-
-    def __contains__(self, key):
-        return self.params and key in self.params
-
-    def copy(self):
-        return self.__class__(self.params)
-
-    @property
-    def query_string(self):
-        """
-        The map request as a query string (the order is not guaranteed).
-
-        >>> qs = RequestParams(dict(foo='egg', bar='ham%eggs', baz=100)).query_string
-        >>> sorted(qs.split('&'))
-        ['bar=ham%25eggs', 'baz=100', 'foo=egg']
-        """
-        kv_pairs = []
-        for key, values in self.params.iteritems():
-            value = ','.join(str(v) for v in values)
-            kv_pairs.append(key + '=' + quote(value.encode('utf-8'), safe=','))
-        return '&'.join(kv_pairs)
-
-    def with_defaults(self, defaults):
-        """
-        Return this MapRequest with all values from `defaults` overwritten.
-        """
-        new = self.copy()
-        for key, value in defaults.params.iteritems():
-            if value != [None]:
-                new.set(key, value, unpack=True)
-        return new
 
 
 class BaseRequest(object):
